requirements.txt
urllib
selenium
PyAutoIt
pyperclip